Add undef protection to categorize
In some cases, messages arrive with message.text set to undefined. This
causes the iSMention check in categorize.js to fail and causes
CoreBot.js to throw a "An error occurred in the categorize middleware:
TypeError: Cannot read property 'indexOf' of undefined" error.

Set the text to "" to allow the checks to occur (and fail) but not
throw an exception.